--- a/packages/dds/matrix/src/test/memory/sharedMatrix.spec.ts
+++ b/packages/dds/matrix/src/test/memory/sharedMatrix.spec.ts
@@ -149,7 +149,7 @@ describe("SharedMatrix memory usage", () => {
 
 							for (let i = 0; i < count; i++) {
 								this.localMatrix.insertCols(Math.floor(this.localMatrix.colCount / 2), 1);
-								this.localMatrix.removeCols(Math.floor(this.localMatrix.rowCount / 2), 1);
+								this.localMatrix.removeCols(Math.floor(this.localMatrix.colCount / 2), 1);
 								this.localMatrix.insertRows(Math.floor(this.localMatrix.rowCount / 2), 1);
 								this.localMatrix.removeRows(Math.floor(this.localMatrix.rowCount / 2), 1);
 							}