## Description

> Remove the previous benchmark tests and add new baseline performance benchmark tests for insert/remove rows, columns, and set cells. Make all the operation in middle and separate insert and remove operation to get better measurement.

## Test result

**Perf test**
~~~
Writing test results relative to package to nyc/junit-report.xml

  SharedMatrix memory usage
    Size of 10*10 SharedMatrix
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 10 times (58ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 100 times (49ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 1000 times (96ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 1000 times (78ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 1000 times (101ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 1000 times (256ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 10 cells
    Size of 100*100 SharedMatrix
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 10 times (84ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 10 times (42ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 100 times (55ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 1000 times (95ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 1000 times (55ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 1000 times (140ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 1000 times (287ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 10 times
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 10 cells
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 100 times
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 100 cells
    Size of 1000*1000 SharedMatrix
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 10 times (613ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 10 times (610ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 10 times (479ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 10 times (561ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 100 times (448ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 100 times (391ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 100 times (428ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 100 times (454ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 1000 times (483ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 1000 times (511ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 1000 times (572ms)
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 1000 times (653ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 10 times (347ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 10 times (349ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 10 times (418ms)
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 10 cells (463ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 100 times (405ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 100 times (474ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 100 times (457ms)
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 100 cells (383ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 1000 times (590ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 1000 times (799ms)
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 1000 times (608ms)
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 1000 cells (532ms)
  60 passing (14s)
~~~

~~~
{
    "suiteName": "SharedMatrix memory usage Size of 10*10 SharedMatrix",
    "benchmarks": [
        {
            "benchmarkName": "Insert a column in the middle 10 times",
            "elapsedSeconds": 30.001050271,
            "customData": {
                "Heap Used Avg": 8655.789888378202,
                "Heap Used StdDev": 5518.139780442991,
                "Margin of Error": 277.1397476095054,
                "Relative Margin of Error": 3.2017845994807512,
                "Iterations": 1603
            }
        },
        {
            "benchmarkName": "Insert a row in the middle 10 times",
            "elapsedSeconds": 30.015023324,
            "customData": {
                "Heap Used Avg": 8628.605566870332,
                "Heap Used StdDev": 5240.640650389919,
                "Margin of Error": 267.6326406603917,
                "Relative Margin of Error": 3.1016905175034477,
                "Iterations": 1551
            }
        },
        {
            "benchmarkName": "Insert a row and a column 10 times",
            "elapsedSeconds": 11.141041328,
            "customData": {
                "Heap Used Avg": 16318.830188679245,
                "Heap Used StdDev": 4999.646288222721,
                "Margin of Error": 388.56778685667643,
                "Relative Margin of Error": 2.3811007429088575,
                "Iterations": 670
            }
        },
        {
            "benchmarkName": "Insert a row and a column and remove them right away 10 times",
            "elapsedSeconds": 30.02536158,
            "customData": {
                "Heap Used Avg": -5050.579345088161,
                "Heap Used StdDev": 60306.459090525146,
                "Margin of Error": 3425.0271630166503,
                "Relative Margin of Error": 67.81454025363628,
                "Iterations": 1254
            }
        },
        {
            "benchmarkName": "Insert a column in the middle 100 times",
            "elapsedSeconds": 11.462548979,
            "customData": {
                "Heap Used Avg": 73483.38364779874,
                "Heap Used StdDev": 23326.31997066504,
                "Margin of Error": 1812.8995540870628,
                "Relative Margin of Error": 2.4670877470424837,
                "Iterations": 670
            }
        },
        {
            "benchmarkName": "Insert a row in the middle 100 times",
            "elapsedSeconds": 9.333959226,
            "customData": {
                "Heap Used Avg": 74479.2183908046,
                "Heap Used StdDev": 21194.728975020436,
                "Margin of Error": 1818.2294479971777,
                "Relative Margin of Error": 2.4412574235898012,
                "Iterations": 550
            }
        },
        {
            "benchmarkName": "Insert a row and a column 100 times",
            "elapsedSeconds": 15.251945884,
            "customData": {
                "Heap Used Avg": 135706.7944358578,
                "Heap Used StdDev": 43983.92199245614,
                "Margin of Error": 3389.2053499495296,
                "Relative Margin of Error": 2.497447061540789,
                "Iterations": 681
            }
        },
        {
            "benchmarkName": "Insert a row and a column and remove them right away 100 times",
            "elapsedSeconds": 30.010890051,
            "customData": {
                "Heap Used Avg": 173546.64732547596,
                "Heap Used StdDev": 121109.78551059656,
                "Margin of Error": 7147.391150858495,
                "Relative Margin of Error": 4.118426521633695,
                "Iterations": 1161
            }
        },
        {
            "benchmarkName": "Insert a column in the middle 1000 times",
            "elapsedSeconds": 1.541951129,
            "customData": {
                "Heap Used Avg": 806617.8723404255,
                "Heap Used StdDev": 530.2198730560126,
                "Margin of Error": 151.58741385964683,
                "Relative Margin of Error": 0.01879296492895843,
                "Iterations": 50
            }
        },
        {
            "benchmarkName": "Insert a row in the middle 1000 times",
            "elapsedSeconds": 1.559633058,
            "customData": {
                "Heap Used Avg": 806540.7659574468,
                "Heap Used StdDev": 396.8455078198327,
                "Margin of Error": 113.45629858327038,
                "Relative Margin of Error": 0.014067025917603322,
                "Iterations": 50
            }
        },
        {
            "benchmarkName": "Insert a row and a column 1000 times",
            "elapsedSeconds": 2.24418278,
            "customData": {
                "Heap Used Avg": 1589488.3404255318,
                "Heap Used StdDev": 403.5045000878173,
                "Margin of Error": 115.36007373035646,
                "Relative Margin of Error": 0.007257686061382035,
                "Iterations": 50
            }
        },
        {
            "benchmarkName": "Insert a row and a column and remove them right away 1000 times",
            "elapsedSeconds": 4.270642632,
            "customData": {
                "Heap Used Avg": 2338751.659574468,
                "Heap Used StdDev": 15769.691741461564,
                "Margin of Error": 4508.48206551371,
                "Relative Margin of Error": 0.19277301405887706,
                "Iterations": 50
            }
        },
        {
            "benchmarkName": "Remove the middle column 10 times",
            "elapsedSeconds": 30.023377151,
            "customData": {
                "Heap Used Avg": -959.0026714158504,
                "Heap Used StdDev": 27137.310612822064,
                "Margin of Error": 1587.2049119234582,
                "Relative Margin of Error": 165.50578629568818,
                "Iterations": 1182
            }
        },
        {
            "benchmarkName": "Remove the middle row 10 times",
            "elapsedSeconds": 4.80450006,
            "customData": {
                "Heap Used Avg": 5049.081081081081,
                "Heap Used StdDev": 959.541058615976,
                "Margin of Error": 126.22435117645112,
                "Relative Margin of Error": 2.4999470032163686,
                "Iterations": 234
            }
        },
        {
            "benchmarkName": "Remove a row and a column 10 times",
            "elapsedSeconds": 30.020783088,
            "customData": {
                "Heap Used Avg": 7630.057613168724,
                "Heap Used StdDev": 6221.157244669613,
                "Margin of Error": 349.81555014352847,
                "Relative Margin of Error": 4.584703915469544,
                "Iterations": 1279
            }
        },
        {
            "benchmarkName": "Set a 3-character string in 10 cells",
            "elapsedSeconds": 30.007798247,
            "customData": {
                "Heap Used Avg": 1439.7415185783523,
                "Heap Used StdDev": 954.6521133200334,
                "Margin of Error": 53.17908827256792,
                "Relative Margin of Error": 3.6936552559155675,
                "Iterations": 1303
            }
        }
    ]
}

~~~

**Local test**
~~~
/home/tong/workspace/FluidFramework/node_modules/.bin/mocha ./sharedMatrix.spec.js --no-timeouts --exit
Writing test results relative to package to nyc/junit-report.xml

  SharedMatrix memory usage
    Size of 10*10 SharedMatrix
      ✔ @Benchmark @Measurement @MemoryUsage Insert a column in the middle 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row in the middle 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Insert a row and a column and remove them right away 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle column 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove the middle row 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Remove a row and a column 5 times
      ✔ @Benchmark @Measurement @MemoryUsage Set a 3-character string in 5 cells
  8 passing (150ms)
~~~