@Josmithr 2025-05-07
> Seems like 10000 times test case does burnout vm's heap. Sample error trace:
> 
> ```
> packages/dds/matrix test:mocha: FATAL ERROR: Ineffective mark-compacts near heap limit Allocation failed - JavaScript heap out of memory
> packages/dds/matrix test:mocha: ----- Native stack trace -----
> packages/dds/matrix test:mocha:  1: 0xb8cf03 node::OOMErrorHandler(char const*, v8::OOMDetails const&) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  2: 0xf04140 v8::Utils::ReportOOMFailure(v8::internal::Isolate*, char const*, v8::OOMDetails const&) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  3: 0xf04427 v8::internal::V8::FatalProcessOutOfMemory(v8::internal::Isolate*, char const*, v8::OOMDetails const&) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  4: 0x1116075  [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  5: 0x1116604 v8::internal::Heap::RecomputeLimits(v8::internal::GarbageCollector) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  6: 0x112d4f4 v8::internal::Heap::PerformGarbageCollection(v8::internal::GarbageCollector, v8::internal::GarbageCollectionReason, char const*) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  7: 0x112dd0c v8::internal::Heap::CollectGarbage(v8::internal::AllocationSpace, v8::internal::GarbageCollectionReason, v8::GCCallbackFlags) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  8: 0x1104011 v8::internal::HeapAllocator::AllocateRawWithLightRetrySlowPath(int, v8::internal::AllocationType, v8::internal::AllocationOrigin, v8::internal::AllocationAlignment) [/usr/local/bin/node]
> packages/dds/matrix test:mocha:  9: 0x11051a5 v8::internal::HeapAllocator::AllocateRawWithRetryOrFailSlowPath(int, v8::internal::AllocationType, v8::internal::AllocationOrigin, v8::internal::AllocationAlignment) [/usr/local/bin/node]
> packages/dds/matrix test:mocha: 10: 0x10e27f6 v8::internal::Factory::NewFillerObject(int, v8::internal::AllocationAlignment, v8::internal::AllocationType, v8::internal::AllocationOrigin) [/usr/local/bin/node]
> packages/dds/matrix test:mocha: 11: 0x153e336 v8::internal::Runtime_AllocateInYoungGeneration(int, unsigned long*, v8::internal::Isolate*) [/usr/local/bin/node]
> ```
> 
> Reduce it to 5000? Ok, 5000 also ran out of heap: https://dev.azure.com/fluidframework/public/_build/results?buildId=336354&view=logs&j=855df218-f14a-5111-4dba-a3c8eab6872c&t=a5446ac8-4e73-5540-487b-79110ad67165
> 
> Just 1000?

I was wondering if this might happen. Reducing to 1000 is fine for now, but we will probably want to investigate further later.