@chentong7 2025-05-12
> A couple of small comments. And we're still bringing back some form of "insert and then remove" tests in this PR, right?

Synced offline, added the insert-and-then-remove cases to ensure memory usage should be very low in this case